+++ src/mdnsDgram.ts
@@ -1,6 +1,7 @@
+/* eslint-disable @typescript-eslint/no-unused-vars */
 /* eslint-disable no-console */
 import dgram from 'dgram';
-import { AnsiLogger, CYAN, db, debugStringify, er, GREEN, hk, idn, ign, LogLevel, nf, rs, TimestampFormat, YELLOW, zb } from 'node-ansi-logger';
+import { AnsiLogger, CYAN, db, debugStringify, er, GREEN, hk, idn, ign, LogLevel, nf, rs, TimestampFormat, wr, YELLOW, zb } from 'node-ansi-logger';
 import os, { NetworkInterfaceInfoIPv4, NetworkInterfaceInfoIPv6 } from 'os';
 import dnsPacket from 'dns-packet';
 
@@ -75,21 +76,14 @@ class MdnsScanner {
     }
   }
 
-  /*
-  [09:42:33.452] [MdnsScanner] Received message from 192.168.1.233:5353
-  [09:42:33.452] [MdnsScanner] Answer: PTR _services._dns-sd._udp.local Data: _http._tcp.local
-  [09:42:33.453] [MdnsScanner] Answer: PTR _http._tcp.local Data: shellybutton1-485519F31EA3._http._tcp.local
-  [09:42:33.453] [MdnsScanner] Answer: SRV shellybutton1-485519F31EA3._http._tcp.local Data: { priority: 0, weight: 0, port: 80, target: 'shellybutton1-485519F31EA3.local' }
-  [09:42:33.454] [MdnsScanner] Answer: A shellybutton1-485519F31EA3.local Data: 192.168.1.233
-  [09:42:33.454] [MdnsScanner] Answer: NSEC shellybutton1-485519F31EA3.local Data: { nextDomain: 'shellybutton1-485519F31EA3.local', rrtypes: [ 'A' ] }   
-ShellyPlus1PM-441793D69718 ShellyPlus1PM-441793D69718
-  */
   private onMessage = (msg: Buffer, rinfo: dgram.RemoteInfo) => {
     // if (this.devices.has(rinfo.address)) return;
     // if (rinfo.address !== '192.168.1.181' && rinfo.address !== '192.168.1.217' && rinfo.address !== '192.168.1.218') return;
+    // if (rinfo.address !== '192.168.1.227' && rinfo.address !== '192.168.1.217' && rinfo.address !== '192.168.1.218') return;
     this.log.info(`Received message from ${ign}${rinfo.address}:${rinfo.port}${rs}`);
     this.devices.set(rinfo.address, { address: rinfo.address, port: rinfo.port });
 
+    /*
     try {
       const decoded = dnsPacket.decode(msg);
       decoded.questions?.forEach((question) => {
@@ -191,9 +185,10 @@ ShellyPlus1PM-441793D69718 ShellyPlus1PM-441793D69718
       this.errorDevices.set(rinfo.address, { address: rinfo.address, port: rinfo.port });
       this.log.error(`Failed to decode mDNS message from ${rinfo.address}:${rinfo.port}: ${error instanceof Error ? error.message : error}`);
     }
+    */
 
     try {
-      this.parseMdnsResponse(msg);
+      this.parseMdnsResponse(msg, rinfo);
     } catch (error) {
       this.errorDevices.set(rinfo.address, { address: rinfo.address, port: rinfo.port });
       this.log.error(`Failed to decode mDNS message from ${rinfo.address}:${rinfo.port}: ${error instanceof Error ? error.message : error}`);
@@ -224,6 +219,18 @@ ShellyPlus1PM-441793D69718 ShellyPlus1PM-441793D69718
   }
 
   private buildQuery(name: string, type: string, classId = 1): Buffer {
+    return dnsPacket.encode({
+      type: 'query',
+      id: 0,
+      flags: 0,
+      questions: [
+        {
+          type: 'PTR',
+          name,
+        },
+      ],
+    });
+    /*
     // Build an mDNS query packet
     const parts = name.split('.').filter((part) => part.length > 0);
     const query = Buffer.alloc(12 + parts.reduce((len, part) => len + part.length + 1, 0) + 5);
@@ -260,6 +267,7 @@ ShellyPlus1PM-441793D69718 ShellyPlus1PM-441793D69718
     query.writeUInt16BE(classId, offset);
 
     return query;
+    */
   }
 
   public stop = () => {
@@ -271,7 +279,7 @@ ShellyPlus1PM-441793D69718 ShellyPlus1PM-441793D69718
     }
   };
 
-  private parseMdnsResponse = (msg: Buffer) => {
+  private parseMdnsResponse = (msg: Buffer, rinfo: dgram.RemoteInfo) => {
     try {
       const id = msg.readUInt16BE(0);
       const flags = msg.readUInt16BE(2);
@@ -345,8 +353,9 @@ ShellyPlus1PM-441793D69718 ShellyPlus1PM-441793D69718
             break;
         }
         console.log(
-          `${name.padEnd(50, ' ')} ${idn}${this.rrTypeToString(type)}${rs}, Class: ${qclass}, TTL: ${ttl}, Data: ${typeof decoded === 'object' ? debugStringify(decoded) : decoded}`,
+          `${name.startsWith('shelly') ? wr : ''}${name.padEnd(50, ' ')} ${idn}${this.rrTypeToString(type)}${rs}, Class: ${qclass}, TTL: ${ttl}, Data: ${typeof decoded === 'object' ? debugStringify(decoded) : decoded}`,
         );
+        if (name.startsWith('shelly')) this.shellyDevices.set(rinfo.address, { address: rinfo.address, id: name.split('.')[0], gen: 0 });
       }
 
       if (nscount > 0) console.log(`${GREEN}Authority RRs:${rs}`);
@@ -393,8 +402,9 @@ ShellyPlus1PM-441793D69718 ShellyPlus1PM-441793D69718
             break;
         }
         console.log(
-          `${name.padEnd(50, ' ')} ${idn}${this.rrTypeToString(type)}${rs}, Class: ${qclass}, TTL: ${ttl}, Data: ${typeof decoded === 'object' ? debugStringify(decoded) : decoded}`,
+          `${name.startsWith('shelly') ? wr : ''}${name.padEnd(50, ' ')} ${idn}${this.rrTypeToString(type)}${rs}, Class: ${qclass}, TTL: ${ttl}, Data: ${typeof decoded === 'object' ? debugStringify(decoded) : decoded}`,
         );
+        if (name.startsWith('shelly')) this.shellyDevices.set(rinfo.address, { address: rinfo.address, id: name.split('.')[0], gen: 0 });
       }
 
       if (arcount > 0) console.log(`${GREEN}Additional RRs:${rs}`);
@@ -441,8 +451,9 @@ ShellyPlus1PM-441793D69718 ShellyPlus1PM-441793D69718
             break;
         }
         console.log(
-          `${name.padEnd(50, ' ')} ${idn}${this.rrTypeToString(type)}${rs}, Class: ${qclass}, TTL: ${ttl}, Data: ${typeof decoded === 'object' ? debugStringify(decoded) : decoded}`,
+          `${name.startsWith('shelly') ? wr : ''}${name.padEnd(50, ' ')} ${idn}${this.rrTypeToString(type)}${rs}, Class: ${qclass}, TTL: ${ttl}, Data: ${typeof decoded === 'object' ? debugStringify(decoded) : decoded}`,
         );
+        if (name.startsWith('shelly')) this.shellyDevices.set(rinfo.address, { address: rinfo.address, id: name.split('.')[0], gen: 0 });
       }
     } catch (error) {
       console.error('Failed to parse mDNS response:', error);
@@ -553,29 +564,20 @@ ShellyPlus1PM-441793D69718 ShellyPlus1PM-441793D69718
     return { nsdname, newOffset };
   }
 
-  // Function to decode NSEC records from an mDNS message
-  // c00c00050000800040 I know this { nextDomain: 'MSL320-bfde._hap._tcp.local', rrtypes: [ 'TXT', 'SRV' ] }
-  /*
-  ID: 0
-  Flags: 33792
-  Questions: 0
-  Answers: 2
-  Authority RRs: 0
-  Additional RRs: 3
-  Answers:
-  _http._tcp.local                                   PTR, Class: 1, TTL: 120, Data: shellyplus1pm-441793d69718._http._tcp.local
-  [19:30:20.200] [MdnsScanner] NSEC rrtypes offset 108 newOffset 110 c04500c02800210001000000780008000000000050c045c028001000010000007800060567656e3d32c04500018001000000780004c0a801d9 00c02800210001000000780008000000000050c045c028001000010000007800060567656e3d32c04500018001000000780004c0a801d9
-  */
   private decodeNSEC(msg: Buffer, offset: number) {
     const { name, newOffset } = this.decodeDomainName(msg, offset);
-    this.log.warn(`***NSEC rrtypes offset ${offset} newOffset ${newOffset}`, msg.toString('hex', offset), msg.toString('hex', newOffset));
+    // this.log.warn(`***NSEC rrtypes offset ${offset} newOffset ${newOffset}`, msg.toString('hex', offset), msg.toString('hex', newOffset));
     offset = newOffset;
 
+    const rrTypes: string[] = [];
     const blockNumber = msg[offset];
     const bitmapLength = msg[offset + 1];
+    if (offset + 2 + bitmapLength > msg.length) {
+      this.log.error(`NSEC bitmap out of bounds`);
+      return { nextDomain: name, rrTypes };
+    }
     const bitmap = msg.slice(offset + 2, offset + 2 + bitmapLength);
 
-    const rrTypes: string[] = [];
     for (let i = 0; i < bitmapLength * 8; i++) {
       if (bitmap[Math.floor(i / 8)] & (0x80 >> i % 8)) {
         const rrType = i + blockNumber * 256;
@@ -730,21 +732,16 @@ ShellyPlus1PM-441793D69718 ShellyPlus1PM-441793D69718
 
 MdnsScanner.listNetworkInterfaces(); // List available network interfaces
 
-const scanner = new MdnsScanner(process.argv[2], false);
+const scanner = new MdnsScanner(process.argv[2], process.argv[3] === 'true' ? true : false);
 
-scanner.query('_shelly._tcp.local', 'PTR', 32769);
 scanner.query('_shelly._tcp.local', 'PTR');
-
-scanner.query('_http._tcp.local', 'PTR', 32769);
 scanner.query('_http._tcp.local', 'PTR');
 // scanner.query('_services._dns-sd._udp.local', 'PTR');
 
-// scanner.query('_services._dns-sd._udp.local', 'PTR', 32769);
-
 const interval = setInterval(() => {
-  scanner.query('_http._tcp.local', 'PTR');
   scanner.query('_shelly._tcp.local', 'PTR');
-  scanner.query('_services._dns-sd._udp.local', 'PTR');
+  scanner.query('_http._tcp.local', 'PTR');
+  // scanner.query('_services._dns-sd._udp.local', 'PTR');
 }, 60000); // Query every 60 seconds
 
 setTimeout(() => {
@@ -752,17 +749,14 @@ setTimeout(() => {
 }, 30000); // Stop scanning after 10 seconds
 
 process.on('SIGINT', () => {
-  clearInterval(interval);
-  scanner.stop();
-
   // Collect devices into an array
   const shellyArray = Array.from(scanner.shellyDevices.entries());
   // Sort the array by device address
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+
   shellyArray.sort(([keyA, deviceA], [keyB, deviceB]) => deviceA.address.localeCompare(deviceB.address));
   // Log the sorted devices
   console.log(`Found ${shellyArray.length} Shelly devices:`);
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+
   shellyArray.forEach(([key, device]) => {
     console.log(`Shelly id: ${hk}${device.id}${rs} host ${zb}${device.address}${rs} gen: ${YELLOW}${device.gen}${rs}`);
   });
@@ -770,11 +764,11 @@ process.on('SIGINT', () => {
   // Collect devices into an array
   const devicesArray = Array.from(scanner.devices.entries());
   // Sort the array by device address
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+
   devicesArray.sort(([keyA, deviceA], [keyB, deviceB]) => deviceA.address.localeCompare(deviceB.address));
   // Log the sorted devices
   console.log(`Found ${devicesArray.length} devices:`);
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+
   devicesArray.forEach(([key, device]) => {
     console.log(`Device: ${device.address}:${device.port}`);
   });
@@ -782,408 +776,15 @@ process.on('SIGINT', () => {
   // Collect devices into an array
   const errorDevicesArray = Array.from(scanner.errorDevices.entries());
   // Sort the array by device address
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+
   errorDevicesArray.sort(([keyA, deviceA], [keyB, deviceB]) => deviceA.address.localeCompare(deviceB.address));
   // Log the sorted devices
   console.error(`Found ${errorDevicesArray.length} devices with errors:`);
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+
   errorDevicesArray.forEach(([key, device]) => {
     console.error(`Device with error: ${device.address}:${device.port}`);
   });
+
+  clearInterval(interval);
+  scanner.stop();
 });
-
-/*
-
-
-
-[21:51:49.227] [MdnsScanner] Received message from 192.168.1.44:5353
-ID: 0
-Flags: 33792
-Questions: 0
-Answers: 1
-Authority RRs: 0
-Additional RRs: 1
-Answers:
-MSL320-bfde._hap._tcp.local                        TXT, Class: 32769, TTL: 4500, Data: c#=2 ff=2 id=0E:65:DD:3F:BA:01 md=MSL320 pv=1.1 s#=1 sf=0 ci=5 sh=pJNf+Q==
-Additional RRs:
-[21:51:49.230] [MdnsScanner] decodeNSEC: jump found at 138 jumpOffset 140 0
-[21:51:49.230] [MdnsScanner] NSEC rrtypes originalOffset 138 jumpOffset 140 offset 41 c00c00050000800040 00050000800040 0010800100001194004b0463233d320466663d321469643d30453a36353a44443a33463a42413a3031096d643d4d534c3332300670763d312e310473233d310473663d300463693d350b73683d704a4e662b513d3dc00c002f8001000011940009c00c00050000800040
-MSL320-bfde._hap._tcp.local                        NSEC, Class: 32769, TTL: 4500, Data: nextDomain MSL320-bfde._hap._tcp.local rrtypes
-
-
-
-
-[21:51:49.231] [MdnsScanner] Received message from fe80::4ae1:e9ff:fe68:bfde%9:5353
-ID: 0
-Flags: 33792
-Questions: 0
-Answers: 1
-Authority RRs: 0
-Additional RRs: 1
-Answers:
-MSL320-bfde._hap._tcp.local                        TXT, Class: 32769, TTL: 4500, Data: c#=2 ff=2 id=0E:65:DD:3F:BA:01 md=MSL320 pv=1.1 s#=1 sf=0 ci=5 sh=pJNf+Q==
-Additional RRs:
-[21:51:49.232] [MdnsScanner] decodeNSEC: jump found at 138 jumpOffset 140 0
-[21:51:49.232] [MdnsScanner] NSEC rrtypes originalOffset 138 jumpOffset 140 offset 41 c00c00050000800040 00050000800040 0010800100001194004b0463233d320466663d321469643d30453a36353a44443a33463a42413a3031096d643d4d534c3332300670763d312e310473233d310473663d300463693d350b73683d704a4e662b513d3dc00c002f8001000011940009c00c00050000800040
-MSL320-bfde._hap._tcp.local                        NSEC, Class: 32769, TTL: 4500, Data: nextDomain MSL320-bfde._hap._tcp.local rrtypes
-
-
-[21:52:35.612] [MdnsScanner] Received message from 192.168.1.12:5353
-ID: 0
-Flags: 33792
-Questions: 0
-Answers: 3
-Authority RRs: 0
-Additional RRs: 1
-Answers:
-Living-room-HomePod-Sx.local                       AAAA, Class: 32769, TTL: 4500, Data: fe80:0000:0000:0000:1c3f:f97c:0122:af6d
-Living-room-HomePod-Sx.local                       AAAA, Class: 32769, TTL: 4500, Data: fd78:cbf8:4939:0746:180d:e446:c47d:68c8
-Living-room-HomePod-Sx.local                       AAAA, Class: 32769, TTL: 4500, Data: fd78:cbf8:4939:0746:0002:fd28:06df:8c5d
-Additional RRs:
-[21:52:35.614] [MdnsScanner] decodeNSEC: jump found at 136 jumpOffset 138 0
-[21:52:35.614] [MdnsScanner] NSEC rrtypes originalOffset 136 jumpOffset 138 offset 42 c00c000440000008 000440000008 001c8001000011940010fe800000000000001c3ff97c0122af6dc00c001c8001000011940010fd78cbf849390746180de446c47d68c8c00c001c8001000011940010fd78cbf8493907460002fd2806df8c5dc00c002f8001000011940008c00c000440000008
-Living-room-HomePod-Sx.local                       NSEC, Class: 32769, TTL: 4500, Data: nextDomain Living-room-HomePod-Sx.local rrtypes
-
-
-
-
-[21:52:35.615] [MdnsScanner] Received message from fe80::1c3f:f97c:122:af6d%9:5353
-ID: 0
-Flags: 33792
-Questions: 0
-Answers: 3
-Authority RRs: 0
-Additional RRs: 1
-Answers:
-Living-room-HomePod-Sx.local                       AAAA, Class: 32769, TTL: 4500, Data: fe80:0000:0000:0000:1c3f:f97c:0122:af6d
-Living-room-HomePod-Sx.local                       AAAA, Class: 32769, TTL: 4500, Data: fd78:cbf8:4939:0746:180d:e446:c47d:68c8
-Living-room-HomePod-Sx.local                       AAAA, Class: 32769, TTL: 4500, Data: fd78:cbf8:4939:0746:0002:fd28:06df:8c5d
-Additional RRs:
-[21:52:35.617] [MdnsScanner] decodeNSEC: jump found at 136 jumpOffset 138 0
-[21:52:35.617] [MdnsScanner] NSEC rrtypes originalOffset 136 jumpOffset 138 offset 42 c00c000440000008 000440000008 001c8001000011940010fe800000000000001c3ff97c0122af6dc00c001c8001000011940010fd78cbf849390746180de446c47d68c8c00c001c8001000011940010fd78cbf8493907460002fd2806df8c5dc00c002f8001000011940008c00c000440000008
-Living-room-HomePod-Sx.local                       NSEC, Class: 32769, TTL: 4500, Data: nextDomain Living-room-HomePod-Sx.local rrtypes
-
-
-
-[19:17:15.075] [MdnsScanner] Failed to decode mDNS message from 192.168.1.218:5353
-[19:17:15.078] [MdnsScanner] Failed to decode mDNS message from 192.168.1.237:5353
-[19:17:15.078] [MdnsScanner] Failed to decode mDNS message from 192.168.1.228:5353
-[19:17:15.079] [MdnsScanner] Failed to decode mDNS message from 192.168.1.229:5353
-[19:17:15.080] [MdnsScanner] Failed to decode mDNS message from 192.168.1.217:5353
-[19:17:15.080] [MdnsScanner] Failed to decode mDNS message from 192.168.1.224:5353
-[19:17:15.081] [MdnsScanner] Failed to decode mDNS message from 192.168.1.239:5353
-[19:17:15.081] [MdnsScanner] Failed to decode mDNS message from 192.168.1.238:5353
-[19:17:18.040] [MdnsScanner] Failed to decode mDNS message from 192.168.1.218:5353
-[19:17:18.041] [MdnsScanner] Failed to decode mDNS message from 192.168.1.228:5353
-[19:17:18.043] [MdnsScanner] Failed to decode mDNS message from 192.168.1.229:5353
-[19:17:18.043] [MdnsScanner] Failed to decode mDNS message from 192.168.1.237:5353
-[19:17:18.044] [MdnsScanner] Failed to decode mDNS message from 192.168.1.239:5353
-[19:17:18.044] [MdnsScanner] Failed to decode mDNS message from 192.168.1.224:5353
-[19:17:18.045] [MdnsScanner] Failed to decode mDNS message from 192.168.1.217:5353
-[19:17:18.045] [MdnsScanner] Failed to decode mDNS message from 192.168.1.238:5353
-[19:18:04.525] [MdnsScanner] Failed to decode mDNS message from 192.168.1.237:5353
-[19:18:04.527] [MdnsScanner] Failed to decode mDNS message from 192.168.1.239:5353
-[19:18:04.528] [MdnsScanner] Failed to decode mDNS message from 192.168.1.238:5353
-[19:18:04.529] [MdnsScanner] Failed to decode mDNS message from 192.168.1.218:5353
-[19:18:04.529] [MdnsScanner] Failed to decode mDNS message from 192.168.1.224:5353
-[19:18:04.530] [MdnsScanner] Failed to decode mDNS message from 192.168.1.229:5353
-[19:18:04.531] [MdnsScanner] Failed to decode mDNS message from 192.168.1.217:5353
-[19:18:04.622] [MdnsScanner] Failed to decode mDNS message from 192.168.1.228:5353
-[19:18:07.497] [MdnsScanner] Failed to decode mDNS message from 192.168.1.239:5353
-[19:18:07.498] [MdnsScanner] Failed to decode mDNS message from 192.168.1.229:5353
-[19:18:07.499] [MdnsScanner] Failed to decode mDNS message from 192.168.1.238:5353
-[19:18:07.500] [MdnsScanner] Failed to decode mDNS message from 192.168.1.218:5353
-[19:18:07.500] [MdnsScanner] Failed to decode mDNS message from 192.168.1.237:5353
-[19:18:07.501] [MdnsScanner] Failed to decode mDNS message from 192.168.1.217:5353
-[19:18:07.502] [MdnsScanner] Failed to decode mDNS message from 192.168.1.224:5353
-[19:18:07.502] [MdnsScanner] Failed to decode mDNS message from 192.168.1.228:5353
-*/
-
-/*
-[17:52:08.021] [MdnsScanner] Received message from 192.168.1.181:5353
-{
-  id: 0,
-  type: 'query',
-  flags: 0,
-  flag_qr: false,
-  opcode: 'QUERY',
-  flag_aa: false,
-  flag_tc: false,
-  flag_rd: false,
-  flag_ra: false,
-  flag_z: false,
-  flag_ad: false,
-  flag_cd: false,
-  rcode: 'NOERROR',
-  questions: [
-    { name: '_shelly._tcp.local', type: 'PTR', class: 'UNKNOWN_32769' }
-  ],
-  answers: [
-    {
-      name: '_shelly._tcp.local',
-      type: 'PTR',
-      ttl: 114,
-      class: 'IN',
-      flush: false,
-      data: 'shellyplus1-e465b8f3028c._shelly._tcp.local'
-    },
-    {
-      name: '_shelly._tcp.local',
-      type: 'PTR',
-      ttl: 114,
-      class: 'IN',
-      flush: false,
-      data: 'shelly1pmmini-348518e04d44._shelly._tcp.local'
-    },
-    {
-      name: '_shelly._tcp.local',
-      type: 'PTR',
-      ttl: 114,
-      class: 'IN',
-      flush: false,
-      data: 'shellyplus2pm-30c92286cb68._shelly._tcp.local'
-    },
-    {
-      name: '_shelly._tcp.local',
-      type: 'PTR',
-      ttl: 114,
-      class: 'IN',
-      flush: false,
-      data: 'shellyplusi4-cc7b5c8aea2c._shelly._tcp.local'
-    },
-    {
-      name: '_shelly._tcp.local',
-      type: 'PTR',
-      ttl: 114,
-      class: 'IN',
-      flush: false,
-      data: 'shellyplus1pm-441793d69718._shelly._tcp.local'
-    },
-    {
-      name: '_shelly._tcp.local',
-      type: 'PTR',
-      ttl: 114,
-      class: 'IN',
-      flush: false,
-      data: 'shelly1mini-348518e0e804._shelly._tcp.local'
-    },
-    {
-      name: '_shelly._tcp.local',
-      type: 'PTR',
-      ttl: 114,
-      class: 'IN',
-      flush: false,
-      data: 'shellyplus2pm-5443b23d81f8._shelly._tcp.local'
-    },
-    {
-      name: '_shelly._tcp.local',
-      type: 'PTR',
-      ttl: 114,
-      class: 'IN',
-      flush: false,
-      data: 'shellyplus2pm-30c922810da0._shelly._tcp.local'
-    },
-    {
-      name: '_shelly._tcp.local',
-      type: 'PTR',
-      ttl: 115,
-      class: 'IN',
-      flush: false,
-      data: 'shellypro2pm-ec62608c9c00._shelly._tcp.local'
-    },
-    {
-      name: '_shelly._tcp.local',
-      type: 'PTR',
-      ttl: 115,
-      class: 'IN',
-      flush: false,
-      data: 'shellypro1pm-ec62608ab9a4._shelly._tcp.local'
-    },
-    {
-      name: '_shelly._tcp.local',
-      type: 'PTR',
-      ttl: 115,
-      class: 'IN',
-      flush: false,
-      data: 'shelly1pmminig3-543204519264._shelly._tcp.local'
-    },
-    {
-      name: '_shelly._tcp.local',
-      type: 'PTR',
-      ttl: 115,
-      class: 'IN',
-      flush: false,
-      data: 'shellypmminig3-84fce63957f4._shelly._tcp.local'
-    },
-    {
-      name: '_shelly._tcp.local',
-      type: 'PTR',
-      ttl: 115,
-      class: 'IN',
-      flush: false,
-      data: 'shelly1minig3-543204547478._shelly._tcp.local'
-    },
-    {
-      name: '_shelly._tcp.local',
-      type: 'PTR',
-      ttl: 115,
-      class: 'IN',
-      flush: false,
-      data: 'shellypro4pm-34987a67d7d0._shelly._tcp.local'
-    }
-  ],
-  authorities: [],
-  additionals: []
-}
-[17:52:08.030] [MdnsScanner] Received message from fe80::2c:2369:caeb:83e2%9:5353
-{
-  id: 0,
-  type: 'query',
-  flags: 0,
-  flag_qr: false,
-  opcode: 'QUERY',
-  flag_aa: false,
-  flag_tc: false,
-  flag_rd: false,
-  flag_ra: false,
-  flag_z: false,
-  flag_ad: false,
-  flag_cd: false,
-  rcode: 'NOERROR',
-  questions: [
-    { name: '_shelly._tcp.local', type: 'PTR', class: 'UNKNOWN_32769' }
-  ],
-  answers: [
-    {
-      name: '_shelly._tcp.local',
-      type: 'PTR',
-      ttl: 114,
-      class: 'IN',
-      flush: false,
-      data: 'shellyplus1-e465b8f3028c._shelly._tcp.local'
-    },
-    {
-      name: '_shelly._tcp.local',
-      type: 'PTR',
-      ttl: 114,
-      class: 'IN',
-      flush: false,
-      data: 'shelly1pmmini-348518e04d44._shelly._tcp.local'
-    },
-    {
-      name: '_shelly._tcp.local',
-      type: 'PTR',
-      ttl: 114,
-      class: 'IN',
-      flush: false,
-      data: 'shellyplus2pm-30c92286cb68._shelly._tcp.local'
-    },
-    {
-      name: '_shelly._tcp.local',
-      type: 'PTR',
-      ttl: 114,
-      class: 'IN',
-      flush: false,
-      data: 'shellyplusi4-cc7b5c8aea2c._shelly._tcp.local'
-    },
-    {
-      name: '_shelly._tcp.local',
-      type: 'PTR',
-      ttl: 114,
-      class: 'IN',
-      flush: false,
-      data: 'shellyplus1pm-441793d69718._shelly._tcp.local'
-    },
-    {
-      name: '_shelly._tcp.local',
-      type: 'PTR',
-      ttl: 114,
-      class: 'IN',
-      flush: false,
-      data: 'shelly1mini-348518e0e804._shelly._tcp.local'
-    },
-    {
-      name: '_shelly._tcp.local',
-      type: 'PTR',
-      ttl: 114,
-      class: 'IN',
-      flush: false,
-      data: 'shellyplus2pm-5443b23d81f8._shelly._tcp.local'
-    },
-    {
-      name: '_shelly._tcp.local',
-      type: 'PTR',
-      ttl: 114,
-      class: 'IN',
-      flush: false,
-      data: 'shellyplus2pm-30c922810da0._shelly._tcp.local'
-    },
-    {
-      name: '_shelly._tcp.local',
-      type: 'PTR',
-      ttl: 115,
-      class: 'IN',
-      flush: false,
-      data: 'shellypro2pm-ec62608c9c00._shelly._tcp.local'
-    },
-    {
-      name: '_shelly._tcp.local',
-      type: 'PTR',
-      ttl: 115,
-      class: 'IN',
-      flush: false,
-      data: 'shellypro1pm-ec62608ab9a4._shelly._tcp.local'
-    },
-    {
-      name: '_shelly._tcp.local',
-      type: 'PTR',
-      ttl: 115,
-      class: 'IN',
-      flush: false,
-      data: 'shelly1pmminig3-543204519264._shelly._tcp.local'
-    },
-    {
-      name: '_shelly._tcp.local',
-      type: 'PTR',
-      ttl: 115,
-      class: 'IN',
-      flush: false,
-      data: 'shellypmminig3-84fce63957f4._shelly._tcp.local'
-    },
-    {
-      name: '_shelly._tcp.local',
-      type: 'PTR',
-      ttl: 115,
-      class: 'IN',
-      flush: false,
-      data: 'shelly1minig3-543204547478._shelly._tcp.local'
-    },
-    {
-      name: '_shelly._tcp.local',
-      type: 'PTR',
-      ttl: 115,
-      class: 'IN',
-      flush: false,
-      data: 'shellypro4pm-34987a67d7d0._shelly._tcp.local'
-    }
-  ],
-  authorities: [],
-  additionals: []
-}
-*/
-/*
-[11:58:27.605] [MdnsScanner] Failed to decode mDNS message from 192.168.1.229:5353: Cannot decode name (buffer overflow) Shelly Plus2PM 30C922810DA0| 1.4.0| ebee9893| 12:08:02 | Feedback
-[11:58:27.605] [MdnsScanner] Failed to decode mDNS message from 192.168.1.238:5353: Cannot decode name (buffer overflow) Shelly Plus1Mini 348518E0E804| 1.4.0| ebee9893| 12:08:21 | Feedback
-[11:58:27.606] [MdnsScanner] Failed to decode mDNS message from 192.168.1.243:5353: Cannot decode name (buffer overflow) Shelly EM Gen3 84FCE636582C| 1.4.99-dev101493| c9daf7c7| 12:08:37 | Feedback
-[11:58:27.607] [MdnsScanner] Failed to decode mDNS message from 192.168.1.228:5353: Cannot decode name (buffer overflow) Shelly Plus2PM 30C92286CB68| 1.4.0| ebee9893| 12:09:23 | Feedback
-[11:58:27.608] [MdnsScanner] Failed to decode mDNS message from 192.168.1.239:5353: Cannot decode name (buffer overflow) Shelly Plus1PMMini 348518E04D44| 1.4.0| ebee9893| 12:09:42 | Feedback
-[11:58:27.608] [MdnsScanner] Failed to decode mDNS message from 192.168.1.218:5353: Cannot decode name (buffer overflow) Shelly Plus2PM 5443B23D81F8| 1.4.0| ebee9893| 12:09:59 | Feedback
-[11:58:27.608] [MdnsScanner] Failed to decode mDNS message from 192.168.1.237:5353: Cannot decode name (buffer overflow) Shelly Plus1 E465B8F3028C| 1.4.0| ebee9893| 12:10:15 | Feedback
-[11:58:27.609] [MdnsScanner] Failed to decode mDNS message from 192.168.1.224:5353: Cannot decode name (buffer overflow) Shelly PlusI4 CC7B5C8AEA2C| 1.4.0| ebee9893| 12:10:29 | Feedback
-[11:58:27.609] [MdnsScanner] Failed to decode mDNS message from 192.168.1.217:5353: Cannot decode name (buffer overflow) Shelly Plus1PM 441793D69718| 1.4.0| ebee9893| 12:10:45 | Feedback
-[11:58:27.609] [MdnsScanner] Failed to decode mDNS message from 192.168.1.227:5353: Cannot decode name (buffer overflow) Shelly Pro1PM EC62608AB9A4| 1.4.0| ebee9893| 12:16:05 | Feedback
-[11:58:27.609] [MdnsScanner] Failed to decode mDNS message from 192.168.1.235:5353: Cannot decode name (buffer overflow) Shelly Pro2PM EC62608C9C00| 1.4.0| ebee9893| 12:16:28 | Feedback
-[11:58:27.609] [MdnsScanner] Failed to decode mDNS message from 192.168.1.242:5353: Cannot decode name (buffer overflow) Shelly DALI Dimmer Gen3 84FCE636832C| 1.4.99-dev101226| b1cdeea9| 12:18:01 | Feedback
-[11:58:27.609] [MdnsScanner] Failed to decode mDNS message from 192.168.1.220:5353: Cannot decode name (buffer overflow) Shelly MiniPMG3 84FCE63957F4| 1.4.0| ebee9893| 12:29:36 | Feedback
-[11:58:27.609] [MdnsScanner] Failed to decode mDNS message from 192.168.1.221:5353: Cannot decode name (buffer overflow) Shelly Mini1G3 543204547478| 1.4.0| ebee9893| 12:30:01 | Feedback
-*/